chore: move popper component
